# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Application",
    "version": "1.0"
  },
  "paths": {
...
    "/enter": {
      "post": {
        "tags": [
          "Entering"
        ],
        "summary": "Создание заявки на поступление",
        "description": "Определяет пользователя по JWT токену и находит его данные. Доступно только Абитуриенту",
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enter/program": {
      "get": {
        "tags": [
          "Entering"
        ],
        "summary": "Получение программ образования с пагинацией и фильтрацией",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Макс. количество объектов на странице",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "faculty",
            "in": "query",
            "description": "Факультет",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "educationLevel",
            "in": "query",
            "description": "Уровень образования",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formOfEducation",
            "in": "query",
            "description": "Форма обучения",
            "schema": {
              "$ref": "#/components/schemas/FormOfEducation"
            }
          },
          {
            "name": "studyingLanguage",
            "in": "query",
            "description": "Язык обучения",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Название программы(по частичному совпадению)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "specialityCode",
            "in": "query",
            "description": "Код специальности",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Entering"
        ],
        "summary": "Изменение приоритета программ внутри заявки",
        "description": "Доступно только абитуриенту. Проверка заявки осуществляется на основе JWT",
        "requestBody": {
          "description": "Список запросов вида programId - приоритет",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EditProgramPriorityRequest"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EditProgramPriorityRequest"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EditProgramPriorityRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponseWithPriority"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponseWithPriority"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponseWithPriority"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enter/program/my": {
      "get": {
        "tags": [
          "Entering"
        ],
        "summary": "Получение программ абитуриента",
        "description": "Доступно только абитуриенту",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/enter/program/{id}": {
      "post": {
        "tags": [
          "Entering"
        ],
        "summary": "Добавление обр. программы в заявку на поступление по id программы",
        "description": "Доступно только абитуриенту",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID программы",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entering"
        ],
        "summary": "Удаление программы из заявки по id",
        "description": "Доступно только абитуриенту. Проверка заявки осуществляется на основе JWT",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID программы",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditProgramPriorityRequest": {
        "required": [
          "priority",
          "programId"
        ],
        "type": "object",
        "properties": {
          "programId": {
            "type": "string",
            "description": "ID программы внутри заявки",
            "format": "uuid"
          },
          "priority": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "description": "Приоритет программы в заявке. Может принимать значения от 1 до N, где N - макс. количество программ в заявке",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Запрос на изменение приоритета программы в заявке абитуриента"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Статус код ошибки",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Ответ на ошибку"
      },
      "FormOfEducation": {
        "enum": [
          "FullTime",
          "Distance",
          "PartTime"
        ],
        "type": "string"
      },
...
      "ProgramResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название программы",
            "nullable": true
          },
          "specialityCode": {
            "type": "string",
            "description": "Код специальности программы (xx.xx.xx)",
            "nullable": true
          },
          "studyingLanguage": {
            "type": "string",
            "description": "Основной язык обучения",
            "nullable": true
          },
          "formOfEducation": {
            "$ref": "#/components/schemas/FormOfEducation"
          },
          "faculty": {
            "type": "string",
            "description": "Название факультета",
            "nullable": true
          },
          "educationLevel": {
            "type": "string",
            "description": "Уровень образования",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Программа обучения"
      },
      "ProgramResponseWithPriority": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название программы",
            "nullable": true
          },
          "specialityCode": {
            "type": "string",
            "description": "Код специальности программы (xx.xx.xx)",
            "nullable": true
          },
          "studyingLanguage": {
            "type": "string",
            "description": "Основной язык обучения",
            "nullable": true
          },
          "formOfEducation": {
            "$ref": "#/components/schemas/FormOfEducation"
          },
          "faculty": {
            "type": "string",
            "description": "Название факультета",
            "nullable": true
          },
          "educationLevel": {
            "type": "string",
            "description": "Уровень образования",
            "nullable": true
          },
          "priority": {
            "type": "integer",
            "description": "Приоритет программы в заявке. Может принимать значения от 1 до N, где N - макс. количество программ в заявке",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Программа с полем \"Приоритет\". Для вывода подробной информации о заявке и её программах"
      },
...
    }
  }
}